+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_07_03_190000) do
+ActiveRecord::Schema[8.1].define(version: 2026_07_04_000002) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
   enable_extension "vector"
@@ -1618,8 +1618,10 @@ ActiveRecord::Schema[8.1].define(version: 2026_07_03_190000) do
     t.string "email_address", null: false
     t.boolean "email_on_mention", default: true, null: false
     t.boolean "email_on_thread_activity", default: true, null: false
+    t.boolean "email_on_waiting_on_replies_digest", default: true, null: false
     t.datetime "email_otp_enabled_at"
     t.jsonb "hidden_calendar_ids", default: [], null: false
+    t.jsonb "inbox_smart_groups", default: {}, null: false
     t.string "locale"
     t.datetime "mfa_last_totp_at"
     t.string "name"